+++ portal/static/portal/js/sticky_subnav.js
@@ -1,28 +1,26 @@
-function toggleStickySubnav (scrollToTop) {
-  $(window).on('scroll', function () {
-    var currentScroll = $(window).scrollTop()
+function toggleStickySubnav(scrollToTop) {
+  $(window).on("scroll", function () {
+    var currentScroll = $(window).scrollTop();
     if (currentScroll >= scrollToTop) {
-      if (!$('.sticky-subnav').hasClass('sub-nav--fixed')) {
-        $('.sticky-subnav').addClass('sub-nav--fixed')
+      if (!$(".sticky-subnav").hasClass("sub-nav--fixed")) {
+        $(".sticky-subnav").addClass("sub-nav--fixed");
       }
-      if (!$('#sticky-warning').hasClass('sub-nav--warning--fixed')) {
-        $('#sticky-warning').addClass('sub-nav--warning--fixed')
+      if (!$("#sticky-warning").hasClass("sub-nav--warning--fixed")) {
+        $("#sticky-warning").addClass("sub-nav--warning--fixed");
       }
-      $('#top').addClass('sub-nav--filler')
+      $("#top").addClass("sub-nav--filler");
     } else {
-      $('#sticky-warning').removeClass('sub-nav--warning--fixed')
-      $('.sticky-subnav').removeClass('sub-nav--fixed')
-      $('#top').removeClass('sub-nav--filler')
-      $('.menu').removeClass('hide')
+      $("#sticky-warning").removeClass("sub-nav--warning--fixed");
+      $(".sticky-subnav").removeClass("sub-nav--fixed");
+      $("#top").removeClass("sub-nav--filler");
+      $(".menu").removeClass("hide");
     }
-  })
+  });
 }
 
 $(function () {
-  $('a.x-icon').on('click', function () {
-    $(this)
-      .parent()
-      .remove()
-    return false
-  })
-})
+  $("a.x-icon").on("click", function () {
+    $(this).parents(".sub-nav").remove();
+    return false;
+  });
+});