+++ portal/tests/cypress/support/commands.ts
@@ -59,11 +59,11 @@ Cypress.Commands.add('loginAsTeacher', () => {
 })
 
 Cypress.Commands.add('loginAsStudent', () => {
-  cy.request(`/login/student/${studentAccessCode}`)
+  cy.request(`/login/student/${studentAccessCode}/`)
   cy.getCookie('csrftoken').then(csrfToken => {
     cy.request({
       method: 'POST',
-      url: `/login/student/${studentAccessCode}`,
+      url: `/login/student/${studentAccessCode}/`,
       failOnStatusCode: true,
       form: true,
       body: {