+++ portal/tests/test_teacher_student.py
@@ -1,5 +1,13 @@
 from __future__ import absolute_import
 
+import json
+import pytest
+
+from django.test import Client
+from django.urls import reverse
+from django.contrib.auth.models import User
+from selenium.webdriver.common.alert import Alert
+
 from common.tests.utils.classes import create_class_directly
 from common.tests.utils.organisation import (
     create_organisation_directly,
@@ -9,6 +17,7 @@ from common.tests.utils.student import (
     create_many_school_students,
     create_school_student,
     create_school_student_directly,
+    create_student_with_direct_login,
 )
 from common.tests.utils.teacher import signup_teacher_directly
 
@@ -20,7 +29,7 @@ class TestTeacherStudent(BaseTest):
     def test_create(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
-        _, _, _ = create_class_directly(email)
+        create_class_directly(email)
 
         self.selenium.get(self.live_server_url)
         page = (
@@ -37,7 +46,7 @@ class TestTeacherStudent(BaseTest):
     def test_create_valid_name_dash(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
-        _, _, _ = create_class_directly(email)
+        create_class_directly(email)
 
         student_name = "Florian-Gilbert"
 
@@ -57,7 +66,7 @@ class TestTeacherStudent(BaseTest):
     def test_create_valid_name_underscore(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
-        _, _, _ = create_class_directly(email)
+        create_class_directly(email)
 
         student_name = "Florian_Gilbert"
 
@@ -77,7 +86,7 @@ class TestTeacherStudent(BaseTest):
     def test_create_invalid_name(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
-        _, _, _ = create_class_directly(email)
+        create_class_directly(email)
 
         student_name = "Florian!"
 
@@ -99,7 +108,7 @@ class TestTeacherStudent(BaseTest):
     def test_create_multiple(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
-        _, _, _ = create_class_directly(email)
+        create_class_directly(email)
 
         self.selenium.get(self.live_server_url)
         page = (
@@ -116,7 +125,7 @@ class TestTeacherStudent(BaseTest):
     def test_create_duplicate(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
-        _, _, _ = create_class_directly(email)
+        create_class_directly(email)
 
         student_name = "bob"
 
@@ -134,6 +143,76 @@ class TestTeacherStudent(BaseTest):
         )
         assert page.adding_students_failed()
         assert page.duplicate_students(student_name)
+
+    def test_onboarding_import_from_csv(self):
+        email, password = signup_teacher_directly()
+        create_organisation_directly(email)
+        create_class_directly(email)
+
+        self.selenium.get(self.live_server_url)
+        page = (
+            HomePage(self.selenium)
+            .go_to_teacher_login_page()
+            .login_no_students(email, password)
+            .import_students_from_csv("test_students_names.csv")
+        )
+
+        assert page.get_students_input_value() == "Student 1\nStudent 2\n"
+
+    def test_onboarding_import_from_csv_error(self):
+        email, password = signup_teacher_directly()
+        create_organisation_directly(email)
+        create_class_directly(email)
+
+        self.selenium.get(self.live_server_url)
+        page = (
+            HomePage(self.selenium)
+            .go_to_teacher_login_page()
+            .login_no_students(email, password)
+            .import_students_from_csv("test_students_names_no_name.csv")
+        )
+
+        alert = Alert(page.browser)
+        assert alert.text == "'Name' column not found in CSV file."
+        alert.dismiss()
+
+    def test_class_students_import_from_csv(self):
+        email, password = signup_teacher_directly()
+        create_organisation_directly(email)
+        _, _, access_code = create_class_directly(email)
+        create_school_student_directly(access_code)
+
+        self.selenium.get(self.live_server_url)
+        page = (
+            HomePage(self.selenium)
+            .go_to_teacher_login_page()
+            .login(email, password)
+            .open_classes_tab()
+            .go_to_class_page()
+            .import_students_from_csv("test_students_names.csv")
+        )
+
+        assert page.get_students_input_value() == "Student 1\nStudent 2\n"
+
+    def test_class_students_import_from_csv_error(self):
+        email, password = signup_teacher_directly()
+        create_organisation_directly(email)
+        _, _, access_code = create_class_directly(email)
+        create_school_student_directly(access_code)
+
+        self.selenium.get(self.live_server_url)
+        page = (
+            HomePage(self.selenium)
+            .go_to_teacher_login_page()
+            .login(email, password)
+            .open_classes_tab()
+            .go_to_class_page()
+            .import_students_from_csv("test_students_names_no_name.csv")
+        )
+
+        alert = Alert(page.browser)
+        assert alert.text == "'Name' column not found in CSV file."
+        alert.dismiss()
 
     def test_add_to_existing_class(self):
         email, password = signup_teacher_directly()
@@ -317,11 +396,11 @@ class TestTeacherStudent(BaseTest):
         assert page.student_exists(name)
         assert page.is_student_password(new_student_password.lower())
 
-    def test_delete(self):
+    def test_delete_student(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
         _, _, access_code = create_class_directly(email)
-        student_name, _, _ = create_school_student_directly(access_code)
+        student_name, _, student = create_school_student_directly(access_code)
 
         self.selenium.get(self.live_server_url)
         page = (
@@ -338,6 +417,50 @@ class TestTeacherStudent(BaseTest):
         page = page.confirm_delete_student_dialog()
 
         assert not page.student_exists(student_name)
+
+        # user/student is removed if they never log in (see below for active student)
+        with pytest.raises(User.DoesNotExist):
+            u = User.objects.get(id=student.new_user.id)
+
+    def test_delete_active_student(self):
+        email, password = signup_teacher_directly()
+        create_organisation_directly(email)
+        _, _, access_code = create_class_directly(email)
+        (
+            student,
+            login_id,
+            student_name,
+            student_password,
+        ) = create_student_with_direct_login(access_code)
+
+        # "active student" is one who has logged in
+        c = Client()
+        url = reverse("student_login", kwargs={"access_code": access_code})
+        c.post(url, {"username": student_name, "password": student_password})
+
+        # teacher login
+        c.login(username=email, password=password)
+
+        # delete the student
+        url = reverse("teacher_delete_students", args=[access_code])
+        data = {"transfer_students": json.dumps([student.id])}
+        c.post(url, data)
+
+        # user should be anonymised
+        u = User.objects.get(id=student.new_user.id)
+        assert u.first_name == "Deleted"
+        assert u.is_active == False
+
+        student.refresh_from_db()
+        assert student.login_id == ""
+
+        # try to login as inactive student, it should fail
+        url = f"/u/{student.user.id}/{login_id}/"
+        response = c.get(url)
+
+        # assert redirects to home page (failed to login)
+        assert response.status_code == 302
+        assert response.url == "/"
 
     def test_reset_passwords(self):
         email, password = signup_teacher_directly()